refactor: split cli into modular components
